test: add a new method of asserting about streams
The stream_recorder module provides a simple way to record and
subsequently assert about the activity of a ReadableStream. Using
pipeTo() to consume a ReadableStream is more reliable than manually
interacting with a reader, as it guarantees that the stream is closed
or errored when the Promise returned from pipeTo() settles.