feat: output stream elements from merge_adjacent_chunks
Previously merge_adjacent_chunks output alternating separator and text
values, representing the merged adjacent separator/text chunks resulting
from partitioning the input raw text chunks.

Now it also performs the additional step of combining separators with
the text that follows them into a single value. And in addition, outputs
values indicating invalid text at the start of a stream, and separators
at the end without subsequent text.

Although this results in the merge step doing more, it has the advantage
that subsequent processing steps need not be context sensitive. Illegal
states such as multiple adjacent separator or text chunks are not
possible, so do not need to be considered downstream.